+++ CHANGELOG.md
@@ -2,9 +2,10 @@
 
 ## [master](https://github.com/arangodb-helper/arangodb/tree/master) (N/A)
 - Fix project CI: linter, tests, build, release
+- Add pass-through options usage hint
 - Return correct exit code when wrong CLI params supplied
 
-## [master](https://github.com/arangodb-helper/arangodb/tree/0.15.4) (2022-03-22)
+## [0.15.4](https://github.com/arangodb-helper/arangodb/tree/0.15.4) (2022-03-22)
 - Use github.com/golang-jwt/jwt
 - Upgrade Go version to 1.17.5